Move `OperationRef` to evc_trace context
