+++ bus-mapping/src/operation.rs
@@ -315,29 +315,6 @@ impl Operation {
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct OperationRef(Target, usize);
-
-impl From<(Target, usize)> for OperationRef {
-    fn from(op_ref_data: (Target, usize)) -> Self {
-        match op_ref_data.0 {
-            Target::Memory => Self(Target::Memory, op_ref_data.1),
-            Target::Stack => Self(Target::Stack, op_ref_data.1),
-            Target::Storage => Self(Target::Storage, op_ref_data.1),
-        }
-    }
-}
-
-impl OperationRef {
-    pub const fn as_usize(&self) -> usize {
-        self.1
-    }
-
-    pub const fn target(&self) -> Target {
-        self.0
-    }
-}
-
 #[cfg(test)]
 mod operation_tests {
     use super::*;
